Rename class HierarchicalIterativeSolver -> solver::HierarchicalIterative.
