+++ CMakeLists.txt
@@ -55,6 +55,7 @@ SET(${PROJECT_NAME}_HEADERS
   include/hpp/constraints/explicit-solver.hh
   include/hpp/constraints/hybrid-solver.hh
   include/hpp/constraints/iterative-solver.hh
+  include/hpp/constraints/solver/hierarchical-iterative.hh
 
   include/hpp/constraints/impl/hybrid-solver.hh
   include/hpp/constraints/impl/iterative-solver.hh
